First working Proof Of Concept
First commit, with a working version of the project.
The creation of the database and following synchronizations seem to be
working, at least without managing conflicts and empty files.

Hashes are not checked, and there is not any way to automize all the
steps, but for now they are just called in a hardcoded way.