Hashes, conflicts, a small service and many small fixes.
Implemented possibility to check hashes, and use them also to try to
solve conflicts when the original file path is not known and multiple
ones match the OneDrive name.
When this does not work, the OneDrive file is simply deleted, then all
the files that conflict are uploaded but OneDrive renames them.
Also, this feature caused some troubles on the upload function, so it
has been modified to.

Added a small script that can be used as a service to keep the content
mirrored, although it is not very advanced.

Some things can be still implemented/improved, e.g. the database
recreation should change the existing flag to 0, and delete items that
do not exist anymore on OneDrive.